requirements.txt
matplotlib
numpy
scipy
sklearn
Pillow
torch
torchvision
torchmetrics
timm